requirements.txt
fastapi
uvicorn
pydantic
pillow
numpy
opencv-python-headless
svgwrite
